refactor(rpc/roadmap): split the roadmap dispatcher into a module directory
rpc/roadmap.rs (2915 lines, nine ops plus tests in one file) becomes
rpc/roadmap/ with one module per responsibility:

  args        arg shapes, decoders, string normalizers
  listing     roadmap_list projection
  intake      roadmap_propose batch validate + insert
  duplicates  near-duplicate title warnings
  deltas      roadmap_propose_update / _discard
  notes       roadmap_note
  brakes      roadmap_hold
  ordering    roadmap_propose_order
  brief       roadmap_brief / _propose_brief_update
  mod.rs      OPS, RoadmapDispatcher, wiring

Tests move to rpc/roadmap/tests/<module>.rs via #[path], with the
shared fixtures in tests/support.rs. Op bodies widen to pub(super) only
where the dispatcher or the fixtures call them. rpc.rs repoints its
#[path] at the new mod.rs.

Frozen: OPS, RoadmapDispatcher::new, every error string and JSON key,
lock-then-emit ordering, test assertions.

cargo test: 1628 passed; fmt and clippy -D warnings clean.
